+++ docusaurus-search-local/src/client/theme/hooks/useSearchQuery.ts
@@ -8,8 +8,12 @@
 import { useHistory, useLocation } from "@docusaurus/router";
 import ExecutionEnvironment from "@docusaurus/ExecutionEnvironment";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
+import {
+  searchContextByPaths,
+} from "../../utils/proxiedGenerated";
 
 const SEARCH_PARAM_QUERY = "q";
+const SEARCH_PARAM_CONTEXT = "ctx";
 
 function useSearchQuery(): any {
   const history = useHistory();
@@ -18,11 +22,13 @@ function useSearchQuery(): any {
     siteConfig: { baseUrl },
   } = useDocusaurusContext();
 
+  const params = ExecutionEnvironment.canUseDOM ? new URLSearchParams(location.search) : null;
+  const searchValue = params?.get(SEARCH_PARAM_QUERY) || "";
+  const searchContext = params?.get(SEARCH_PARAM_CONTEXT) || "";
+
   return {
-    searchValue:
-      (ExecutionEnvironment.canUseDOM &&
-        new URLSearchParams(location.search).get(SEARCH_PARAM_QUERY)) ||
-      "",
+    searchValue,
+    searchContext,
     updateSearchPath: (searchValue: string) => {
       const searchParams = new URLSearchParams(location.search);
 
@@ -37,8 +43,12 @@ function useSearchQuery(): any {
       });
     },
     generateSearchPageLink: (searchValue: string) => {
+      const searchParams = new URLSearchParams(location.search);
+      const searchContext = searchParams.get(SEARCH_PARAM_CONTEXT) || "";
       // Refer to https://github.com/facebook/docusaurus/pull/2838
-      return `${baseUrl}search?q=${encodeURIComponent(searchValue)}`;
+      return `${baseUrl}search?q=${encodeURIComponent(searchValue)}${
+        Array.isArray(searchContextByPaths) ? `&ctx=${encodeURIComponent(searchContext)}` : ""
+      }`;
     },
   };
 }