+++ docusaurus-search-local/src/declarations.ts
@@ -17,8 +17,11 @@ declare module "*/generated.js" {
   export const explicitSearchResultPath: boolean;
   export const searchBarShortcut: boolean;
   export const searchBarShortcutHint: boolean;
+  export const searchBarPosition: "left" | "right";
   export const docsPluginIdForPreferredVersion: string;
   export const indexDocs: boolean;
+  export const searchContextByPaths: string[];
+  export const hideSearchBarWithNoSearchContext: boolean;
   // These below are for mocking only.
   export const __setLanguage: (value: string[]) => void;
   export const __setRemoveDefaultStopWordFilter: (value: boolean) => void;