+++ docusaurus-search-local/src/server/utils/processPluginOptions.ts
@@ -1,9 +1,16 @@
 import path from "path";
-import { PluginOptions, ProcessedPluginOptions } from "../../shared/interfaces";
+import {
+  DocusaurusContext,
+  PluginOptions,
+  ProcessedPluginOptions,
+} from "../../shared/interfaces";
 
 export function processPluginOptions(
   options: PluginOptions | undefined,
-  siteDir: string
+  {
+    siteDir,
+    siteConfig: { themeConfig },
+  }: Pick<DocusaurusContext, "siteDir" | "siteConfig">
 ): ProcessedPluginOptions {
   const config = { ...options } as ProcessedPluginOptions;
   ensureArray(config, "docsRouteBasePath");
@@ -20,6 +27,13 @@ export function processPluginOptions(
   );
   config.docsDir = config.docsDir.map((dir) => path.resolve(siteDir, dir));
   config.blogDir = config.blogDir.map((dir) => path.resolve(siteDir, dir));
+  if (config.searchBarPosition === "auto") {
+    const search = themeConfig.navbar?.items?.find(
+      (item) => item.type === "search"
+    );
+    config.searchBarPosition =
+      search && search.position === "left" ? "left" : "right";
+  }
   return config;
 }