+++ docusaurus-search-local/src/server/utils/validateOptions.ts
@@ -38,7 +38,12 @@ const schema = Joi.object<PluginOptions>({
   ignoreFiles: isArrayOfStringsOrRegExpsOrStringOrRegExp.default([]),
   searchBarShortcut: Joi.boolean().default(true),
   searchBarShortcutHint: Joi.boolean().default(true),
+  searchBarPosition: Joi.string().default("auto"),
   docsPluginIdForPreferredVersion: Joi.string(),
+  zhUserDict: Joi.string(),
+  zhUserDictPath: Joi.string(),
+  searchContextByPaths: Joi.array().items(Joi.string()),
+  hideSearchBarWithNoSearchContext: Joi.boolean().default(false),
 });
 
 export function validateOptions({